// frontend/src/app/model/SinistreFront.ts
import { Credit } from "./Credit";
import { Mvtsinistre } from "./Mvtsinistre";
import { RisqueLocatif } from "./RisqueLocatif";
import { Sinistre } from "./Sinistre";

export class SinistreFront {

    sinistreForm: Sinistre;
    mvtsinistreForm: Mvtsinistre;
    creditForm: Credit;
    risque_locatifForm: RisqueLocatif;
}