// frontend/src/app/model/User.ts
export class User {
    util_id: String;
    util_numero: String;
    util_nom: String;
    util_prenom: String;
    util_denomination: String;
    util_sigle: String;
    util_type: String;
    util_adresse: String;
    util_telephonefixe: String;
    util_telephoneportable: String;
    util_email: String;
    util_numclient: String;
    util_direction: String;
    util_departement: String;
    util_service: String;
    util_poste: String;
    util_status: String;
    util_login: String;
    util_password: String;
    util_profil: String;

}